The dtype of `paddings` and `mask` is now `jnp.bool`. (#1133)
Previously, it was typically `int32`, but since `paddings` only store 0/1
values, that's redundant. We prefer to use `jnp.bool`, which uses only 1 byte.
For reference, paddings are currently used inconsistently across the codebase;
some places use int32, others float32, and elsewhere bool. This PR standardizes
all of them to bool.

Additionally, **XLA may store booleans as 1 bit per value on TPU via
bit-packing** (though this is not officially documented).

When switching paddings from `int32` to `bool`, the following improvements in
memory and runtime were observed:
- **HBM usage**: 86,242.00 MiB → 86,232.91 MiB
- **Step time** (measured 3 times): 1,345,702,414 ± 128,038 ns → 1,345,565,230 ± 136,309 ns

* int32 paddings:
![image](https://github.com/user-attachments/assets/937fac0e-f5b5-4ef9-a6d6-766c083e75c8)

* bool paddings:
![image](https://github.com/user-attachments/assets/69bae5dd-5db6-43cf-9a7d-178c44caaf69)

+++ axlearn/audio/decoder_asr.py
@@ -498,7 +498,7 @@ class CTCDecoderModel(BaseASRDecoderModel):
         max_decode_len = paddings.shape[-1] + 1
         beam_search_outputs = beam_search_decode(
             inputs=jnp.zeros_like(paddings),
-            time_step=jnp.zeros(paddings.shape[0], dtype=paddings.dtype),
+            time_step=jnp.zeros(paddings.shape[0], dtype=jnp.int32),
             cache={"time_step": jnp.array(0)},
             tokens_to_scores=self._tokens_to_scores(input_batch, num_decodes=num_decodes),
             num_decodes=num_decodes,
@@ -539,7 +539,7 @@ class CTCDecoderModel(BaseASRDecoderModel):
         max_decode_len = paddings.shape[-1] + 1
         sample_decode_outputs = sample_decode(
             inputs=jnp.zeros_like(paddings),
-            time_step=jnp.zeros(paddings.shape[0], dtype=paddings.dtype),
+            time_step=jnp.zeros(paddings.shape[0], dtype=jnp.int32),
             cache={"time_step": jnp.array(0)},
             tokens_to_scores=self._tokens_to_scores(
                 input_batch, num_decodes=num_decodes, logits_modifier=logits_modifier
